+++ modules/log-time/controllerV2.js
@@ -11,7 +11,7 @@ const mongoose = require("mongoose")
 const fs = require("fs");
 const { updateUserFun } = require('../usersModule/controller');
 const { myCache } = require('../../Config/config');
-const { updateProjectInternal } = require('../Project/controller/updateProject');
+const { updateProjectInternal } = require('../project/controller/updateProject');
 const loggerConfig = require('../../Config/loggerConfig');
 const socketEmitter = require('../../event/socketEventEmitter.js');
 const { handleFileUploadForTrackerSS,handleuploadMainFileForbase64Thumbnail } = require(`../../common-storage/common-${process.env.STORAGE_TYPE}.js`);
