+++ modules/project/controller/manageGlobalFilter.js
