+++ modules/project/controller/tags.js
